requirements.txt
# Core runtime dependencies
openai-whisper
deep-translator>=1.9

# Development / testing
pytest>=7
pytest-cov
ruff
